squashfile: Initial git repository build
I've reached what I consider the MVP for parsing squashfs files: I can
correctly parse an entire ubuntu rootfs packed into an archive. There
are still a few limitations:

- There are likely a ton of bugs. My first priority will be rooting
  these out and establishing a comprehensive test suite (with code
  coverage metrics!).

- I haven't done any benchmarking/optimization. "Make it work, make it
  right, make it fast".

- Compression is not yet supported. This seems anathema to the
  squashfs format, but building archives with "-noI -noId -noD -noF
  -noX" is sufficient for my purposes. I'll add this in a future
  release.

- Extended attributes are not yet supported. This is a bit lower
  priority, but I'll get to it eventually.

I'm not bothering to pull in the full history here---it is
uninteresting and fairly useless as documentation, being cluttered
with too many "wip", "format", and "fix it" commits.

Let it rip!

Signed-off-by: Brian Kubisiak <brian@kubisiak.com>