squashfile: Fix incorrect id table compression check
In squashfs-tools, there are notes that uncompressed inodes implies
that the id table is likewise uncompressed, regardless of the '-noId'
flag. The linux kernel does not seem to handle the 'uncompressed id
table' flag at all (presumably this information is pulled from the
metadata headers instead of flags).

Change the flag check here to match the behavior of
squashfs-tools. That is, only indicate that id tables are compressed
if the inodes are compressed as well. Notably, this flag should not be
used internally and the individual metadata headers should be used
instead.

Signed-off-by: Brian Kubisiak <brian@kubisiak.com>