squashfile: Silence DirIndex compiler warnings
The DirIndex is used to speed up searching for files in extended
directory inodes. It provides an index for searching for files at
specific offsets within the directory without having to scan through
every file.

My main use case is for scanning through every file in an archive, so
speeding up these lookups is not useful right now. If this changes in
the future, I may use the DirIndex.

Signed-off-by: Brian Kubisiak <brian@kubisiak.com>