squashfile: Remove 'sparse' field
Extended file inodes specify the number of bytes saved by making the
file sparse. This is used in the linux kernel for sparse file
accounting, but is not necessary for parsing the archive
itself. Remove this unused field from the inode structure to save
space and silence a compiler warning.

Signed-off-by: Brian Kubisiak <brian@kubisiak.com>