squashfile: Silence unused lookup table field
Squashfs archives may optionally contain an "export table" for quickly
mapping between inode numbers and on-disk inode locations (for NFS
exports). Mark this as unused in the superblock.

This may be useful in the future for faster iteration over the inodes,
with the downside that it won't be able to look up paths.

Signed-off-by: Brian Kubisiak <brian@kubisiak.com>