squashinfo: Add method for retrieving inode number
The inode number is a unique, opaque reference to each entry in the
squashfs archive. There's not a strict 1-1 mapping between inode
number and squashinfo due to hard links, but this may still be useful
for fast lookup tables etc.

In the future, this may also be used to index into the export table,
if this ends up being useful.

Signed-off-by: Brian Kubisiak <brian@kubisiak.com>